avformat/demux: Add more retries to get more stream durations
The number of extra retries to get more than a single duration is currently limited to 1 (see commit 77a0df4b5).
This patch raises this number of retries to 4 (amongst the overall 6 max retries).
In my experience, the need for 2 extra retries is something commonplace.
3 extra retries is not far away and we have an exemple on ffmpeg.org :
http://samples.ffmpeg.org/archive/extension/ts/mpegts+mpeg2video+ac3++mpegts_multiple_ts_packets_pmt_pid_0x50.ts
And I have encountered at least one time a file (SPTS) that requires 4 extra retries.